Fixed: Websocket Session save into redis
