+++ api/src/main/java/com/pfplaybackend/api/config/websocket/domain/entity/domainmodel/PartyroomSession.java
@@ -1,4 +1,4 @@
-package com.pfplaybackend.api.partyroom.domain.entity.domainmodel;
+package com.pfplaybackend.api.config.websocket.domain.entity.domainmodel;
 
 import com.pfplaybackend.api.partyroom.domain.value.PartyroomId;
 import com.pfplaybackend.api.user.domain.value.UserId;
@@ -9,14 +9,12 @@ import lombok.Setter;
 @Setter
 @Getter
 public class PartyroomSession {
-    private long id;
     private String sessionId;
     private UserId userId;
     private PartyroomId partyroomId;
 
     @Builder
-    public PartyroomSession(long id, String sessionId, UserId userId, PartyroomId partyroomId) {
-        this.id = id;
+    public PartyroomSession(String sessionId, UserId userId, PartyroomId partyroomId) {
         this.sessionId = sessionId;
         this.userId = userId;
         this.partyroomId = partyroomId;
