+++ api/src/main/java/com/pfplaybackend/api/config/websocket/event/SubscriptionEventListener.java
@@ -2,7 +2,7 @@ package com.pfplaybackend.api.config.websocket.event;
 
 import com.pfplaybackend.api.config.jwt.dto.UserCredentials;
 import com.pfplaybackend.api.config.oauth2.dto.CustomAuthentication;
-import com.pfplaybackend.api.partyroom.application.dto.PartyroomSessionDto;
+import com.pfplaybackend.api.config.websocket.application.service.SessionEventService;
 import com.pfplaybackend.api.partyroom.application.service.PartyroomInfoService;
 import com.pfplaybackend.api.partyroom.domain.value.PartyroomId;
 import com.pfplaybackend.api.partyroom.exception.InvalidPartymemberException;
@@ -11,7 +11,6 @@ import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.context.ApplicationListener;
-import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.messaging.simp.stomp.StompHeaderAccessor;
 import org.springframework.stereotype.Component;
 import org.springframework.web.socket.messaging.SessionSubscribeEvent;
@@ -24,8 +23,8 @@ import java.util.UUID;
 public class SubscriptionEventListener implements ApplicationListener<SessionSubscribeEvent> {
     //
     private final static Logger logger = LoggerFactory.getLogger(SubscriptionEventListener.class);
+    private final SessionEventService sessionEventService;
     private final PartyroomInfoService partyroomInfoService;
-    private final RedisTemplate<String, Object> redisTemplate;
 
     @Override
     public void onApplicationEvent(SessionSubscribeEvent event) {
@@ -34,27 +33,32 @@ public class SubscriptionEventListener implements ApplicationListener<SessionSub
 
         CustomAuthentication authentication = (CustomAuthentication) headerAccessor.getUser();
         if (authentication == null) {
-            logger.error("Uid: null" + " Session ID: " + sessionId + " Partyroom ID: null");
+            logger.error("Invalid Partymember, UserId is null" + " Session ID: " + sessionId);
             throw new InvalidPartymemberException();
         }
 
         UserCredentials userCredentials = (UserCredentials) authentication.getPrincipal();
         UUID uid = userCredentials.getUid();
         UserId userId = UserId.create(uid);
+
         Optional<PartyroomId> optional = partyroomInfoService.getPartyroomId(userId);
         if (optional.isPresent()) {
             PartyroomId partyroomId = optional.get();
-            PartyroomSessionDto partyroomSessionDto = partyroomInfoService.saveSession(sessionId, uid, partyroomId);
+            sessionEventService.saveSession(sessionId, userId, partyroomId);
             logger.info(
-                    "Uid: " + partyroomSessionDto.getUserId().getUid().toString()
-                            + " Session ID: " + partyroomSessionDto.getSessionId()
-                            + " Partyroom ID: " + partyroomSessionDto.getPartyroomId().getId()
+                    "Session ID: " + sessionId
+                            + " UserID: " + uid
+                            + " Partyroom ID: " + partyroomId
                             + "has subscribed"
             );
             return;
         }
 
-        logger.error("Uid: " + uid.toString() + " Session ID: " + sessionId + " Partyroom ID: " + optional);
+        logger.error(
+                "Invalid Partymember, PartyroomId is null, "
+                        + "Session ID: " + sessionId
+                        + " UserID: " + uid
+        );
         throw new InvalidPartymemberException();
     }
 }
