--- api/src/main/java/com/pfplaybackend/api/partyroom/application/dto/PartyroomSessionDto.java
@@ -1,14 +1,0 @@
-package com.pfplaybackend.api.partyroom.application.dto;
-
-import com.pfplaybackend.api.partyroom.domain.value.PartyroomId;
-import com.pfplaybackend.api.user.domain.value.UserId;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class PartyroomSessionDto {
-    private String sessionId;
-    private UserId userId;
-    private PartyroomId partyroomId;
-}
