+++ api/src/main/java/com/pfplaybackend/api/partyroom/repository/custom/PartyroomRepositoryCustom.java
@@ -1,7 +1,6 @@
 package com.pfplaybackend.api.partyroom.repository.custom;
 
 import com.pfplaybackend.api.partyroom.application.dto.*;
-import com.pfplaybackend.api.partyroom.domain.entity.data.PartyroomSessionData;
 import com.pfplaybackend.api.user.domain.value.UserId;
 
 import java.util.List;
@@ -13,5 +12,4 @@ public interface PartyroomRepositoryCustom {
     Optional<ActivePartyroomWithMemberDto> getMyActivePartyroomWithMemberIdByUserId(UserId userId);
     List<PartyroomWithMemberDto> getMemberDataByPartyroomId();
     Optional<PartyroomIdDto> getPartyroomDataWithUserId(UserId userId);
-    PartyroomSessionDto savePartyroomSession(PartyroomSessionDto partyroomSessionDto);
 }
